procesador-stream/requirements.txt
apache-flink==1.18.0
psycopg2-binary
google-generativeai
spacy